+++ ElementoMapa.java
@@ -16,13 +16,13 @@ public interface ElementoMapa {
     // Retorna a mensagem de interação do elemento
     String interage();
 
-    int getX();
+    double getX();
 
-    int getY();
+    double getY();
 
-    void setX(int x);
+    void setX(double x);
 
-    void setY(int y);
+    void setY(double y);
 
     int getQuantidadeVida();
