+++ Vida.java
@@ -4,8 +4,8 @@ public class Vida implements ElementoMapa {
     private Color cor;
     private Color corPisca;
     private String simbolo;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private boolean status;
     private final int quantidadeVida = 10;
 
@@ -16,19 +16,19 @@ public class Vida implements ElementoMapa {
         this.status = true;
     }
 
-    public int getX(){
+    public double getX(){
         return x;
     }
 
-    public int getY(){
+    public double getY(){
         return y;
     }
 
-    public void setX(int x){
+    public void setX(double x){
         this.x = x;
     }
 
-    public void setY(int y){
+    public void setY(double y){
         this.y = y;
     }
